+++ lib/common/constant.dart
@@ -11,6 +11,7 @@ import 'package:flutter/material.dart';
 
 const appName = 'Panorama Secure Access';
 const appHelperService = 'PSGHelperService';
+const legacyAppHelperService = 'FlClashHelperService';
 const coreName = 'clash.meta';
 const browserUa =
     'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36';